requirements.txt
Django==2.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
pytz==2018.7
django-rest-auth==0.9.3
django-allauth==0.39.1

